requirements.txt
Flask==3.1.2
flask-cors==6.0.1
Flask-Limiter==3.5.0
moviepy==2.1.2
yt-dlp==2025.9.26
Pillow==10.4.0
numpy==2.3.3
werkzeug==3.1.3
gunicorn==21.2.0
psycopg2-binary==2.9.10
redis==5.2.1
rq==2.1.0

# Scaling dependencies
celery==5.4.0
kombu==5.4.2
gevent==24.2.1
eventlet==0.36.1

# Performance monitoring (optional)
flower==2.0.1
prometheus-flask-exporter==0.23.0